# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      summary: List team members
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      operationId: createUser
      tags:
        - users
      summary: Create team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      operationId: getUser
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      operationId: updateUser
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      operationId: deleteUser
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
  /tasks:
    get:
      operationId: listTasks
      tags:
        - tasks
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      operationId: createTask
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    get:
      operationId: getTask
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    patch:
      operationId: updateTask
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      operationId: deleteTask
      tags:
        - tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
  /tickets:
    get:
      operationId: listTickets
      tags:
        - tickets
      summary: List support tickets
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
    post:
      operationId: createTicket
      tags:
        - tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /tickets/{id}:
    get:
      operationId: getTicket
      tags:
        - tickets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
    patch:
      operationId: updateTicket
      tags:
        - tickets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
    delete:
      operationId: deleteTicket
      tags:
        - tickets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          description: super_admin | manager | employee | client
        status:
          type: string
          description: active | suspended
        avatarUrl:
          type:
            - string
            - 'null'
        projectsCount:
          type:
            - integer
            - 'null'
        createdAt:
          type: string
        isActive:
          type: boolean
        roleId:
          type: integer
    UserInput:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
        roleId:
          type: integer
...
    Task:
      type: object
      required:
        - id
        - title
        - status
        - priority
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type:
            - string
            - 'null'
        projectId:
          type:
            - integer
            - 'null'
        projectName:
          type:
            - string
            - 'null'
        assigneeId:
          type:
            - integer
            - 'null'
        assigneeName:
          type:
            - string
            - 'null'
        status:
          type: string
          description: todo | in_progress | review | done
        priority:
          type: string
          description: low | medium | high | urgent
        dueDate:
          type:
            - string
            - 'null'
        createdAt:
          type: string
    TaskInput:
      type: object
      required:
        - title
        - status
        - priority
      properties:
        title:
          type: string
        description:
          type: string
        projectId:
          type: integer
        assigneeId:
          type: integer
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        projectId:
          type: integer
        assigneeId:
          type: integer
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
...
    Ticket:
      type: object
      required:
        - id
        - subject
        - priority
        - status
        - createdAt
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        clientId:
          type:
            - integer
            - 'null'
        clientName:
          type:
            - string
            - 'null'
        subject:
          type: string
        description:
          type:
            - string
            - 'null'
        priority:
          type: string
          description: low | medium | high | urgent
        status:
          type: string
          description: open | in_progress | resolved | closed
        assigneeId:
          type:
            - integer
            - 'null'
        assigneeName:
          type:
            - string
            - 'null'
        createdAt:
          type: string
    TicketInput:
      type: object
      required:
        - subject
        - priority
        - status
      properties:
        clientId:
          type: integer
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        assigneeId:
          type: integer
    TicketUpdate:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        assigneeId:
          type: integer
...