# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  status?: string;
  avatarUrl?: string;
  isActive?: boolean;
  roleId?: number;
}